requirements.txt
pymongo~=4.10.1
streamlit~=1.40.2
sentence-transformers~=4.0.2
pandas~=2.2.3
emoji~=2.14.1
nltk~=3.9.1
google-genai
google
python-dotenv~=1.0.1
protobuf~=3.20.0
requests~=2.32.3
beautifulsoup4~=4.12.3
urllib3~=2.2.3
genai~=2.0.0
numpy~=1.26.4